Combine to 1 IntepreterSpec.
